Move common step definitions
Step definitions that don't relate to a particular feature folder with
feature files have been moved to the top level common folder so that
it's easier to see which ones are global and which ones are specific
to a particular feature.
